remove poc sink
