+++ nusamai/src/lib.rs
@@ -18,5 +18,4 @@ pub static BUILTIN_SINKS: &[&dyn sink::DataSinkProvider] = &[
     &sink::noop::NoopSinkProvider {},
     &sink::minecraft::MinecraftSinkProvider {},
     &sink::obj::ObjSinkProvider {},
-    &sink::cesiumtiles_poc::CesiumTilesPocSinkProvider {},
 ];
