--- nusamai/src/sink/cesiumtiles_poc/tiling/mod.rs
@@ -1,5 +1,0 @@
-pub mod scheme;
-pub mod tree;
-
-pub use scheme::*;
-pub use tree::*;
